Bug#29374606: MEMORY LEAK IN TCP_SOCKET::GET_LISTENER_SOCKET()
Problem: Running on a system where ipv6 address resolution works, but
ipv6 socket creation fails would lead to a memory leak.

Root cause was that the failure to create the ipv6 socket would trigger
a fallback to ipv4 address resolution, and the resulting addrinfo
pointer was assigned to the variable holding the ipv6 addrinfo pointer
without deallocating the referenced object first.

Solution: Use a std::unique_ptr<> to own the addrinfo object,
and use std::unique_ptr<>::operator=() when switching to the ipv4 version.

Change-Id: If76dc99305877e8077e27b373de77f659c95bcf7