Bug#30031130: MYSQL SERVER SEGFAULTS AT STD::CONDITION_VARIABLE::WAIT_FOR IN SOLARIS/X86
When a MySQL Server built with ndbcluster engine is run in a
Solaris/x86 platform, it crashes everytime when
NDB_SCHEMA_OBJECT::client_wait_completed() method is called. The
segfault happens when this method calls the
std::condition_variable::wait_for() method. This particular crash
happens only in the release builds. The reason seems to be an issue with
the compiler used in Solaris. When the optimization level -xO2 is used,
it fills the wait_for() call with a bogus code that triggers this
segfault during runtime.

Fixed the issue by making the release builds use optimization
level -xO1 for ndbcluster instead.

Change-Id: I8471970169cd971958ac2f5aa6dcb0ccd8d083a3