Bug#29911440 DISCOVER SHOULD NOT COMMIT TRANSACTION
Problem
=======
When altering a table in the middle of an transaction, schema distribution
distributes the changes to all the mysqld and because of the change in the
table structure, mysqld doing the ongoing transaction will force a table
discover in the middle of the transaction. The table discovery triggers a
commit and will cause the transaction to be committed.

Fix
===
The ongoing transaction is aborted since an attempt to discover the table
means that the table definition has changed since the beginning of the
transaction.

Change-Id: I3812bc29c3620499598a438e6434f07ed8cbde43