Bug#30017509 RECENT REGRESSION: EXISTS CRASHING IN FLATTEN_SUBQUERIES
The fix for Bug 29759277 added this to a loop in flatten_subqueries():
    if (!cond_value) {
       (*subq)->sj_selection = Item_exists_subselect::SJ_ALWAYS_FALSE;
+      (*subq)->walk(&Item::clean_up_after_removal...
to eliminate dangling pointers.
However, a side effect of clean_up_after_removal (since WL 9571)
is that it removes the subquery from the SELECT_LEX::sj_candidates array,
which we're precisely iterating on at the moment.
Thus, right after the cleaning up, 'subq' is actually pointing to the
next subquery, while we think it's the removed subquery and treating
it as such (first problem). Second problem: the incrementing
at the beginning of the next iteration of the loop makes us jump over
that subquery (not prepare it for semijoin).

Fix:
- first problem: cache *subq into a new local variable 'subq_item'
(and, for symmetry, do it in other similar loops too)
- second problem: decrement subq to cancel the loop's incrementation.
- also renamed subq_pred (which sounded like "the subquery predicate" i.e.
the Item_in_subselect) to "subq_where" (which is closer to what it really is).

Approved by: Roy Lyseng <roy.lyseng@oracle.com>

Change-Id: If71402d0d2ff6e1ce1827907ef817460ab3222a0