Bug#28948915 DEFAULT ROLE IS NOT LOGGED INTO THE BINARY LOG
Description:
------------
Right before committing the ACL statement, we rewrite the ACL statement for
binary log. We missed to write the 'DEFAULT ROLE' clause in the rewritten
CREATE USER statement. Hence, this information was getting missed in the
binary log.

Fix:
---
We were already adding 'default role' clause in the rewritten the query for
SHOW CREATE USER statement. We now use the same method to rewirte both the
CREATE USER and SHOW CREATE USER statements.

Test:
-----
 - Added tests in the rpl.rpl_roles.test file
 - PB2 testing
    - push id: 15036507
    - 2019-07-11 13:08:23

Review:
-------
RB#22500