Bug #29692250 WRONG RETURNING VALUE AND ASSERTION IN FIL_SHARD::OPEN_FILE WHICH CAUSE CRASH
Problem
--------
Server hit a debug assert while attempting to open a file, if the disk is full.
This can happen only in the test environment when the server is forced to throw
a "disk full" error while opening a file readonly.

Solution
---------
Removed the assert. This allows the server to retry opening the file and throw a
fatal error if it cannot be done within a predefined number of retries.

RB# 22568
Reviewed by: Kevin Lewis (kevin.lewis@oracle.com)