BUG#30035857
Fixed a bug introduced by fix for BUG#29506869.
Problem was that now LCPs are executed very early so they can start even before
DICT has committed the table and thus we can abort the CREATE TABLE statement.
In this case we need to wait for LCP to finish, but we still need to
perform the removal of the table.

Reviewed-by: Maitrayi Sabaratnam <maitrayi.sabaratnam@oracle.com>