Bug#30051303: SERVER CRASHED WHILE RUNNING INSERT ON DUPLICATE KEY UPDATE OVER VIEW
Fix an issue where setting up a VALUES table alias in an INSERT .. ODKU
statement did not properly check if the fields inserted into are of type
Item_view_ref, which leads to an erroneous down_cast to Item_field.

In the case of such an INSERT .. ODKU statement where we are inserting
into specific columns of a view (e.g. `INSERT INTO v(a, b)`), the
corresponding Item_field must first be extracted from its reference.
Thus add a check and dereference to setup_values_table in the case of
view insertion.

Change-Id: If1957e6aae5b728a55de83dd2298ca8d96748ffa