Bug#29697588: MYSQLD DEADLOCK OF ALL CLIENT THREADS ORIGINATING FROM 3-WAY               DEADLOCK
Problem
-------
MySQL server can enter into a deadlock when below threads are executed
in parallel.

1. Server thread executing START SLAVE SQL_THREAD.
2. Stopping SQL Thread.
3. Server thread calling show_heartbeat_period() function
   through any of the following queries

   -  SHOW STATUS LIKE 'Slave_heartbeat_period'
   -  SELECT * FROM INFORMATION_SCHEMA.SESSION_STATUS
   -  SELECT * FROM INFORMATION_SCHEMA.GLOBAL_STATUS

Analysis
--------
This is a classic example of deadlock of three threads wherein each
thread holds one lock and waits for another lock in a cyclic manner.

The locks involved in the deadlock are as follows.
+--------------------------------------+-------------+
|                 Lock                 | Scope/Layer |
+--------------------------------------+-------------+
| Multisource_info::m_channel_map_lock | Replication |
+--------------------------------------+-------------+
| Relay_log_info::run_lock             | Replication |
+--------------------------------------+-------------+
| LOCK_status                          | SQL         |
+--------------------------------------+-------------+


Below is the deadlock scenario,

1) START SLAVE SQL_THREAD
   Owns: Multisource_info::m_channel_map_lock
   Waits for: Relay_log_info::run_lock

2) Stopping SQL thread
   Owns: Relay_log_info::run_lock
   Waits for: LOCK_status

3) SHOW STATUS LIKE 'Slave_heartbeat_period'
   Owns: LOCK_status
   Waits for: Multisource_info::m_channel_map_lock


Thread 1 is executing START SLAVE SQL THREAD

  In start_slave_cmd(), the thread acquires channel_map.wrlock(), and
  calls start_slave_thread() to start SQL thread; where it also
  acquires Relay_log_info::run_lock, and calls mysql_thread_create()
  to create a new thread to execute handle_slave_sql() and then enters
  into mysql_cond_wait() waiting for SQL thread to start. After it
  returns from the mysql_cond_wait() on receiving signal from SQL
  thread, it unlocks the Relay_log_info::run_lock temporarily to
  restore to the previous state by calling THD::exit_cond(). This
  unlock allows the thread 2 to acquire the Relay_log_info::run_lock
  and in turn causes the reacquisition of Relay_log_info::run_lock to
  hang.

Thread 2 is replication SQL-thread which is terminating its work.

  Thread 2 is started by thread 1. Thread 2 starts its execution from
  handle_slave_sql(), encounters with some error and jumps to the err
  section where in the end, it acquires Relay_log_info::run_lock, and
  calls THD::release_resources() where it tries to acquire
  LOCK_status.

Thread 3 is a server thread executes something like SHOW STATUS LIKE
'Slave_heartbeat_period'.

  In fill_status(), it acquires LOCK_status mutex, and then calls,
  show_heartbead_period() where in turn it tries to acquire
  channel_map.rdlock().

So thread 1 waits for Relay_log_info::run_lock, which is owned by
thread 2, for LOCK_status, which is owned by thread 3, which waits for
channel_map.rdlock(), which is owned by thread 1.

Fix
---
The three locks mentioned above can be logically divided into two
groups.

- Group 1 consists of Multisource_info::m_channel_map_lock and
  Relay_log_info::run_lock. Both these locks are replication layer
  locks and the order of locking should always be
  Multisource_info::m_channel_map_lock followed by
  Relay_log_info::run_lock. Please see rpl_slave.h for order of
  acquisition for replication locks.

- Group 2 consists of LOCK_status which is a SQL layer lock.

In the above problematic example, when seen from each thread
individually, it seems that there is no problem in the way the locks
are being acquired as there is no specific predefined lock order to
follow in the MySQL system when it comes to locks across these two
groups.

Since there is no problem in the order of acquisition of the locks,
this patch aims at only changing the duration of the locks in thread 2
to break the deadlock.

Before the fix, thread 2(stopping SQL thread) held the
Relay_log_info::run_lock for the complete duration until the cleanup
and the destruction of SQL thread took place.

With this fix, SQL thread is made to release Relay_log_info::run_lock
before calling THD::release_resources and this fixes the deadlock
cycle reported in the bug.

This patch also changes order of the broadcast and mutex unlock calls
at two places in the handle_slave_sql() function in order to follow
the general practice of broadcasting/signaling while holding the
mutex.

Reviewed-by: Pedro Figueiredo <pedro.figueiredo@oracle.com@oracle.com>
Reviewed-by: Abhinav Agarwal <abhinav.ab.agarwal@oracle.com>

RB:22414