Bug#29895213 CLUSTERJ FAILS TO LOAD FROM AN EXTERNAL JAVA MODULE
A web application can be deployed as multiple modules in a web server.
When ClusterJ jar is deployed as such a module, so that multiple web
applications can use it to access NDB, it throws the following exception
everytime an application tries to create a new instance:

java.lang.IllegalArgumentException: non-public interface is not defined by
the given loader

When the application requests a new domain object, ClusterJ tries to
create a proxy class from which the object can be instantiated.
ClusterJ defines the proxy class as an implementation of the domain
interface and the protected DomainTypeHandlerImpl::Finalizable interface.
The class loader of these two interfaces will be different as they belong
to different modules running in the Web Server. When ClusterJ tries to
create the proxy class using domain object interface's class loader,
the above said exception is thrown as it cannot access the definition
of the protected Finalizable interface from another module.

Fixed this issue by making the Finalizable interface public so that the
ClassLoader of the web application would be able to access it even if
they belong to different modules.

Change-Id: Ia78d42f7486c37b18c7288885d2f61f5364e5e1a