Bug#29596969: DEFAULT EXPRESSION IS IGNORED IN CREATE TABLE QUERY, IF THE FIELD IS MARKED AS PK
When setting a column as PRIMARY KEY, there's a check if there is a
DEFAULT value provided by the user, either as autoincrement, a constant
value or as CURRENT_TIMESTAMP and if neither is true, the DEFAULT is set
as null.
The fix is to also include DEFAULTs generated from expressions in the
check in order to skip setting the flag to a wrong value.

Change-Id: I4b38de1917a7eec6c07651f15a5a4b698fd50a28