Bug #29557747: DEBUG-ASSERT WHILE RESTARTING SERVER POST INSTALL COMPONENT RB#22585
The server was trying to binlog "special tables" creation, i.e. the
tables that plugins create that are not exactly tables but rather just
a way for DD to get a grip about tables existing in plugins, but not yet
in DD.
No need to be doing this since this is part of the initialization of all
plugins/components.
Fixed by turning off binlog for that operation.
Test case added.