Bug#29511118 set_persist_only allows to set a role with system_user privilege
Description:
------------
Check() method specified for the system variable is not called for
SET PERSIST_ONLY option. That is the reason SET PERSIST_ONLY statement succeeds
in this case. However, server restart used to fail with the appropriate error.
This behavior could be problematic in situations when user does not have
access to modify the cnf files directly.

Hence, now the behavior will be -
- Continute to throw error in case of SET GLOBAL/PERSIST mandatory_roles =
'system_user_role'
- Continue to execute SET PERSIST_ONLY mandatory_roles = 'system_user_roles'
- Let the server restart with a warning in case a mandatory_roles is set with a
role that has SYSTEM_USER privilege.

Fix:
----
In case the server is starting and invalid role is found in the mandatory_roles
then throw the warning in the server log ans reset the LEX_STRING to empty
value. Discarding the existing LEX_STRING is safe since it is allocated in
the mem_root of the THD or argv_alloc mem_root(if read from cnf file).

Test:
----
Added more tests to verify the scenarios.
mysql-8.0-stage run in progress `
push id: 15031307
Time stamp: 2019-07-10 14:13:54

Review:
RB#22552