BUG#30080194
The list c_backups contains both backup records and LCP records.
There was several misunderstandings of what is first in the list and
what is in the list. Changed to use get_lcp_record to get LCP record
and to use get_backup_record to get backup record. get_backup_record
can return false indicating that no backup record currently exists.
Fixes a number of issues related to this problem.

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>