Bug#30079649 UPGRADE TO GOOGLETEST VERSION 1.8.1
Upgrade to the latest version of googletest, to get better support for
modern C++ features.

On Solaris: patch googletest source for a known compiler bug.
Also patch for bad feature test in gtest-port.h: we *do* have __GLIBCXX__
defined, even if this is not gcc.

Remove all usage of GTEST_HAS_PARAM_TEST, as this is no longer maintained
(all compilers support it)

Change-Id: I803bc3851e8a9527090b347224c466b5f1911948