WL#13029 ENABLE UPGRADE AND DOWNGRADE OF SAME PROCESS
Modifies atrt to support upgrade and downgrade of the same process.

Initially, upgrade/downgrade process involved three steps namely,
stopping the process, switching configuration and starting the process.
Switching back to the original version wasn't possible. Now, atrt is
enabled upgrade/downgrade the processes to the original version and if
the cluster is changed back to the version it was before the execution
of the testcase, cluster is not restarted to start the new testcase.
Cluster is only restarted only if atrt detects a change in the
configuration of the cluster which is inidicated by a new flag, m_changed.
The configuration of cluster processess is stored in a variable called
m_save with a flag (m_saved) indicating the same. But, while setting up
the testcase configuration of API process is saved so that regardless
the result of the test, testcase configuration should be restored to
original. This makes it necessary to have another flag to check whether
the configuration of a process is changed or not.

Now, ndb_api processes (testcase itself) are also restarted in the same
way as the cluster processes, but currently we don't have a testcase
which restarts itself to an earlier version and a special parameter or
an argument needs to be implemented in order to avoid running into an
infinite loop of upgrades and downgrades.So, support for ndb_api
processes will be added later.

Change-Id: I7478f8654c5d40216ed439d234f2fa10d240a3ee