Bug#30090080 AUTO SYNC OF TABLESPACES FAILS TO INSTALL DATAFILES
Problem:
--------

Tablespaces and datafiles are not tightly coupled in NDB in the
sense that they are represented by independent NDB Dictionary
objects. Thus, when metadata is restored using the ndb_restore
tool, there's no guarantee that the tablespace and its associated
datafile objects are restored at the same time. This leads to the
possibility that the tablespace mismatch is detected and
automatically synchronized to DD before the datafile is restored
to NDB. This holds good for logfile groups and undofiles as well

Fix:
----

The metadata change monitor now submits tablespaces and logfile
groups only if their corresponding datafiles and undofiles exist
in NDB

Change-Id: I70702e9cf0c96b242a02feb88b71a1e76c0a6c4c