Bug #29882690	UNDETECTED DEADLOCK WITH GAP AND INSERT INTENTION LOCKS IS POSSIBLE
When fixing Bug #29882690 the deadlock detection mechanism was (re)moved from the main
control path, and accidentally the fragment:

  if (err != DB_DEADLOCK) {
    lock_update_age(lock, m_rec_id.m_heap_no);
  }

was removed, instead of being replaced with an unconditional:

  lock_update_age(lock, m_rec_id.m_heap_no);

This post-push fix, brings the call to lock_update_age() back.
It also removes some MTR scenarios which unfortunately fail in this
implementation of CATS, which can not handle deferred deadlock detection.
Updating CATS algorithm will be addressed in a separate bug.

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
RB:22652