BUG#30086352: LOCAL SYSFILE NOT UPDATED BETWEEN FIRST LCP DONE AND PHASE 50
Reviewed-by: Frazer Clement <frazer.clement@oracle.com>

The local sysfile wasn't updated after first LCP done when new GCIs was saved.
If the restart crashed before phase 50 but after first LCP was done it would
cause an assert to fail that checked the validity of the restorable GCI. To
uphold this assert we needed to write the GCIs in the local sysfile also
after first LCP until phase 49 arrived when the local sysfile is updated with
the state restart completed.