Bug#29900671: mysqld should ignore special . prefixed snapshot directories during startup
Any directory or file that starts with '.' will be considered HIDDEN on any OS including
Windows so that tablespaces can be portable.
HIDDEN files and directories are not scanned at startup and are thus not recovered.
But it is still possible to use HIDDEN directories in DDL.

Approved by Marcin in RB#22606