Bug#30068072: INNODB: PREVENT FILE DISCOVERY FOR FILES WITH HIDDEN AND SYSTEM ATTRIBUTES
This is a continuation of Bug#29900671 which allowed InnoDB to ignore
directories and files that start with "." when scanning at startup
for file discovery. This bug completes the task for Windows where
HIDDEN and SYSTEM directories are identified by attributes.  File
attributes of SYSTEM will also considered HIDDEN from file discovery.

HIDDEN files and directories are not scanned at startup and are thus not
recovered. But it is still possible to use HIDDEN directories in DDL.

Approved by Marcin in RB#22699