WL#12717 SUPPORT DIFFERENT FAILURE MODES IN ATRT
Modifies atrt to accept another option for a test case and a new
command line argument for it's default value to support different failure
modes on test failure.

A new option called "on-failure" is added for the test case and a command
line argument called "--default-behavior-on-failure" is also added to
support 4 different failure modes namely "Restart", "Continue", "Skip",
and "Abort". When a test case fails, the decision about whether to continue
executing the test suite or not is dependent on this optional parameter.
This value is read through an enum variable which has values of type ulong
since enum values are handled as ulong type internally in MySQL
libraries and a custom typelib is provided for the set of values.

• If the option is "Skip", We mark the rest of the test cases as "Skipped"
• If the option is "Abort", We bail out running the testsuite.
• If the option is "Restart", We restart the cluster for the next testcase.
• If the option is "Continue", We continue with the execution of rest of
   the test cases assuming the cluster is alright.

E.g:-
cmd: testBasic
args: -n UnlockBatch T6
max-time: 180
on-failure: <Restart>/<Abort>/<Skip>/<Continue>

Note: These options are case sensitive.

Change-Id: Ie7e54cc7868ed440d7b9bf364fc769f8b1ec28d9