Bug #27484514 NDB : API NODES QUITE NOISY WAITING TO CONNECT WHILE CLUSTER STARTS
Reviewed-by: Frazer Clement <frazer.clement@oracle.com>

1) In the node-restart case where data nodes have started but
not yet elected a President, the mgmd gets a 'node ID already
in use' error. This results in excessive retries and noise. This
is fixed by introducing a new 'not ready to alloc nodeID' error
for the case where nodes have started but not elected a President.

2) In the node-restart case where data nodes have not yet completed
node failure handling, a 'Failed to alloc nodeID' error is returned.
This is fixed by adding a check to detect an incomplete node start
and return a 'Node failure handling not completed'.

3) The retry frequency is reduced for 'not ready to alloc nodeID'
errors.

4) An error insert is added  to simulate a slow restart for testing.

5) Log messages are reworded to indicate that the nodeID alloc errors
are minor and temporary

Change-Id: Icf63078f31e3e3626ceaeb536e5effd5a1cf5a9d