Bug#29915241 "EXPLAIN" REGRESSION IN TEST GCOL_KEYS_NDB
In tests gcol_keys_ndb and gcol_select_ndb, set ndb_index_stat_enable=OFF
and re-record the result files.

NDB's index statistics are built from sampling one fragment of a table. Each
fragment is likely to be statistically representative if the number of rows
in the table is large relative to the number of fragments, but not if it is
small. The fragment chosen for sampling depends on the table's internal ID
number. When seemingly unrelated metadata changes cause this ID to change,
a different fragment may be sampled, resulting in quite different statistics.

For tests that do ANALYZE and EXPLAIN on small tables, disable NDB index
statistics so that the results of EXPLAIN do not change unpredictably
over time.

Reviewed-By: Magnus Blaudd <magnus.blaudd@oracle.com>
Change-Id: I4ce45a18da76b4e75ca3301bcbed5e408b429633