Bug#30083125: ASSERT IN MDL_KEY::MDL_KEY_INIT() AT SQL/MDL.H
Regression from Bug no 27482976: CONSISTENTLY CONSTRUCT
TABLE_LIST OBJECTS. In that patch, two TABLE_LIST
constructors were combined in an unfortunate way. In one
scenario, a TABLE_LIST object is created from a TABLE
representing a temporary table. In the other scenario,
another TABLE_LIST is created for a base table, where the
new one apparently has a longer life span that the original
TABLE_LIST, and hence it has to copy the strings for
database, table, and alias names. For some reason,
TABLE_LIST does not own these strings.

The problem manifests itself in the first scenario,
temporary tables.  In the old constructor, the table name
was set to the same as the alias. In the new one, it is
copied from the TABLE object. In the case of temporary
tables, however, the table name is a file path, which may
exceed the limit for MDL_key names, causing failed
assertion.

Fixed by reintroducing dedicated constructors, behaving the
way the old code did.

We can't write an MTR test for this, as the bug depends on
the length of the path, so we we add a unit test instead.

Change-Id: Ibe2778d6ffe6abe77b3fb4fad992b8296f8f7887