Bug #30086559	CATS ASSERTION FAILURE: NEW_LOCK->TRX->STATE == TRX_STATE_ACTIVE
Post-push-fix:
There are call paths in which we call lock_update_age() without holding trx->mutex,
which means that the thread which runs the trx is allowed to move to COMMITTED_IN_MEMORY
state. It can not progress further, because that would require lock_sys->mutex which
we hold during lock_update_age(), so the trx can not finish the commit nor be freed,
so accessing it's fields (like trx->age) is safe.
Note that since the first implementation we access and modify trx->age field without
holding trx->mutex (see for example lock_update_trx_age()) relaying on the fact that
we hold lock_sys->mutex, and know that trx is not freed yet.

This patch removes the assert entirely.

Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>
RB:22662