Bug #29952775 SETTING SORT_BUFFER_SIZE TO A LARGE VALUE CAUSES QUERY TO GO OUT OF MEMORY
The range optimizer decides to use index merge, and creates a Unique
instance which does init_tree() with sort_buffer_size/16 =
1152921504606846975 as allocation size for its associated
memroot. This fails at the first insert into the tree because malloc
returns nullptr.

To avoid instant out-of-memory when inserting into the tree, we
allocate a tree with default allocation size, rather than sortbuff_size.

Change-Id: I2f129cdfb6d6e0a4aa0905c19eb1e58a74538071