Bug#30114226: THERE IS AN OBVIOUS ERROR IN THE FUNCTION TRX_RSEG_ADD_ROLLBACK_SEGMENTS
Fixed this error as suggested.

This number is used to do a checkpoint if INNODB_ROLLBACK_SEGMENTS is
increased at runtime.  That probably only happens during ntest.
It is not very likely in the field because the default value is the
maximum value of 128 and there is no good reason to ever make it less
than 128.  But this bug could explain some errors found only in ntest.

Approved by Rahul Agarkar in RB#22736