Bug#28268680: REGRESSION: SLOW_QUERIES STATUS VARIABLE ISN'T INCREMENTED IF SLOW QUERY IS OFF
Despite of the manual entry on Slow_queries (the number of
queries that have taken more than long_query_time seconds)
stating that "this counter increments regardless of whether
the slow query log is enabled", the value was only incremented
when the slow log was turned on. With this patch, the behavior
is as per the docs, and the counter is increased whether or not
the slow log is enabled, as long as the query was slow.

Approved by: Jens Even Blomsoy <jens.even.blomsoy@oracle.com>