Bug#30133062: HA_EXAMPLE.SO IS NOT BUILT ON 5.6
This is a backport of Bug#17772560.

The -DWITH_EXAMPLE_STORAGE_ENGINE=1 CMake option was being ignored.
It has now been fixed.

Also, if -DWITH_EXAMPLE_STORAGE_ENGINE=0 is given, the EXAMPLE storage
engine is built as a plugin.

This patch also updates DEB and RPM packaging for ha_example.so.

Reviewed-by: Tor Didriksen <tor.didriksen@oracle.com>