Bug#30096843 MYSQLROUTER FAILS TO START IN MACOS
After unpacking the router packet on macos and successful bootstrap
running the router fails with:

Error: dlopen(.../metadata_cache.so, 6):
 Library not loaded: @loader_path/../lib/libssl.1.0.0.dylib
  Referenced from: .../metadata_cache.so
  Reason: image not found

The problem is that for the router plugins we are using the function
SET_PATH_TO_SSL that assumes that the binary that it
is called for is on the same directory level as the lib directory
(like inside lib or bin dirs).
That is not the case for the router plugins as these are installed
inside lib/mysqlrouter

Solution: setting the path 'manually' to @loader_path/ and creating symbolic
links to ssl libraries in the same dir where the modules that require it.

RB: 22702