Bug#29870919 INFORMATION SCHEMA STATS EXPIRY RESULTS IN BAD STATS FOR PARTITIONED TABLES
INFORMATION_SCHEMA in 8.0 is optimized to fetch table dynamic
statistics from SE, without expecting users to run ANALYZE TABLE
command. An exception to this (for now) is that the dynamic table
statistics for partitioned tables are not optimized.

The problem reported here is that, we request dynamic index
cardinality value using SE API get_index_column_cardinality()
which is not designed to work with partitioned table.  This
patch avoids the same. We open the partitioned table and find
cardinality.

Change-Id: I6b112b81b063db23b82ea71f1cb9238b6b9d3311