Bug #29008298 MYSQLD CRASHES ITSELF WHEN CREATING INDEX
Problem
-------
A long running ALTER TABLE ADD INDEX with concurrent inserts causes sempahore waits and
eventually crashes the server.

Solution
---------
The indexes on the intermediate tables are built using bulk load insert. A concurrent DML
at this stage does not acquire the index lock. An index lock acquired on the intermediate
table, which is not visible to anyone else, does not block concurrent DMLs.

RB# 22482
Reviewed by: Annamalai Gurusami (annamalai.gurusami@oracle.com)