Bug#30158625 MAIN.MAX_STATEMENT_TIME: QUERY EXECUTION WAS INTERRUPTED
The test sets a max execution time of 1 millisec to abort a sleep that
lasts for 5 seconds. If the sleep is interrupted, it returns 1.

However, since the max execution time is very small in various places
in the test (1 msec), the statement might be aborted before the sleep
even starts. This makes the test fail.

Stabilize the test by setting max execution time to a much higher
value to make sure the sleep statment is started properly before it
is interrupted.

In addition, fix the test for the relative precedence of setting max
execution time at session level and at statement level, which was
obviously incorrect.

Change-Id: I82627de03600bdd89fdf3081f5d587afe199bb4c