BUG#30071270: FIX CODE MESSAGES FOR DISTRIBUTED RECOVERY AND CLONE
The recovery process you can see log messages that read like we do
distributed recovery or we clone. But the truth is that clone is a part
of distributed recovery, not an alternative.

The messages should not refer clone as a external part of distributed
recovery but as provisioning step of it.

ReviewBoard: 22786