Bug #30110851: SUBQUERY INVOLVES COUNT() AGGREGATE FUNCTION PERFORMANCE REGRESSION
When determining whether to recalculate a materialized derived table for each
execution, the uncacheability of the _parent_ query block was used instead of
the rematerialize flag on the table (which is currently only true for
JSON_TABLE). This could cause unneeded rematerializations, especially when making
queries against non-merged views, leading to reduced performance.

Do note that there are situations where we do the opposite (materialize where a
derived table or CTE is only used once, so it should be streamed); a TODO has
been added.

(count() has nothing to do with it; the bug title is misleading.)

Change-Id: I2b0f20d2323b4f089f67bd0d0c29361156c97190