BUG#29543447: ALTER TABLE ALLOWS TDE ENCRYPTION ON MYISAM ENGINE
Post-push fix:

The fix for this bug rejects even the ENCRYPTION clause value 'n/N' and not
just 'y/Y'. It is enough to reject 'y/Y' for the fix. Otherwise, There is
possibility that a table created in older version of MySQL to have
ENCRYPTION='n' and a ALTER TABLE could fail after it is upgraded.

This patch allows values 'n/N' with ENCRYPTION clause for non-InnoDB table.
This is inline with behaviour of MySQL 8.0 and higher versions.

In order to test the ENCRYPTION='y', the patch moves test to new test file
innodb.table_encryption.

Change-Id: I41467194ba45f7f9970d6350c41bd92ea02f13ae