WL#12915 - Thread pool optimizations.
The worklog increases the max limit of thread pool size
from 64 to 512. It introduces a system vairable
thread_pool_max_active_query_threads. This is a per group setting
that limits the maximum number of active threads in a thread
group. This is useful when the workload has lot of stalls and
value of 2/3 is optimal for this system variable. The system
variable default value is 0 which shall use the current mechanism in
thread pool of limiting the number of active threads in a thread
group. In addition the thread creation rate is throttled if the
the maximum number of threads in thread pool exceeds a certain
threshold and there are active threads (which may soon return to
the worker pool and pick some query for processing).

Reviewed-by: Dmitry Shulga <dmitry.shulga@oracle.com>