WL#13277 REPLACE T18 WITH T1 IN AUTOTEST TESTCASES
Modifies testScan to use table T1 instead of T18 and removes T18 from
the set of tables in NDBT_Tables.cpp.

testScan requires a table with just a single column primary key to check
for corrupted tuples. So, it doesn't need a special table with two
column primary key, T18. And, T18 isn't being used by any of the
testcases other than testScan. A test was done by going back to the
commit which introduced the testcase to make sure that it can use an
existing table T1 instead of creating it's own table. So, testScan is
modified to use T1 and the definition of T18 is removed from
NDBT_Tables.cpp.

Change-Id: I78474415d0ec00466b0f25cecfe2cbbd3e3a55c0