Bug#30040815: SERVER STARTUP ON WIN FAILS IF DATA_HOME_DIR AND DATADIR USES DIFF LETTER CASES
In order to add a case insensitive filepath comparison on case insensitive
file systems, the previous method of consolidating the list of scan
directories using std::sort() and std::unique is replaced with
Tablespace_dirs::add_path() that compares a new scan directory with
each scan directory currently in the list. This provides more control
and allows more checks.

This change is done for all OSes but only affects OSes that use case
insensitive file systems, typically Windows and OSX, because it adds a
new check for is_file_system_case_insensitive() and if so, checks if the
directory to scan or an ancestor of it is already in the list using a
case insensitive comparison called innobase_nocase_compare()

Some changes were made in related code to improve comments, report
more cases where scan directories are superfluous, and other improvements.

Approved by Marcin in RB#22834