Bug#30068110: AUTOMATIC UPGRADE DOES NOT RELOAD AUDIT PLUGIN
When upgrading the mysqld server to 8.0, we have two main use cases;
upgrading from 5.7 and upgrading from 8.0:

From 5.7: First, the DD tables and I_S view are created, then the
          dynamic plugins are initialized, then the DD meta data is
          migrated from the 5.7 format to the 8.0 format, followed
          by the creation of the P_S tables, and finally, the SQL
          scripts altering the remaining system tables are executed.

From 8.0: First, the DD tables are upgraded, including the meta
          data they contain, then the plugins are initialized,
          then the I_S views and P_S tables are updated, and finally,
          the SQL scripts altering the remaining system tables are
          executed.

For both use cases, when the plugins are initialized, they may not
see the environment they rely on, if they e.g. expect certain mysql
tables to be present with a certain definition, as in the case of
the audit_log plugin. For upgrade from 5.7, this may be because the
tables' meta data has not yet been migrated into the 8.0 format, but
it may also be due to the SQL script upgrading the tables not being
executed yet. For upgrades from 8.0, the latter may cause the plugin
to see an outdated environment when it is initialized. For the
audit_log plugin, this may cause the plugin to start in legacy mode
where certain functionality is disabled.

Restarting the server, or re-installing the plugin, is a workaround
that will cause the plugin to be initialized correctly.

This patch provides a short term solution by delaying initializing
of plugins where this is an issue to the end of upgrade. The two
plugins 'audit_log' and 'mysql_firewall' need to be handled this way.

Change-Id: I28a38ded891badb15da6128db554676ddfb0b4c8