Bug #30191394: ALTERNATIVEITERATOR DOES NOT RESET THE HANDLER WHEN SWITCHING
When switching from table scan to index lookups or the opposite,
AlternativeIterator does not reset the handler, causing assertion failures.

This is rather hard to trigger currently, since most of the paths that would
take AlternativeIterator still run in the pre-iterator executor. It will trigger
in e.g. main.subquery_antijoin when we convert the index subquery engines
to the iterator executor.

Change-Id: Iaf978464d2cdd523098675ecc76dc51c4e767e8e