Bug#29597896 - NULL POINTER DEREFERENCE IN LIBMYSQL                (CLIENT.CC FUNCTION UNPACK_FIELD)
DESCRIPTION:
============
There is a null pointer dereference in libmysql file
in function unpack_field while fetching client fields
metadata sent by the server. This results in a crash
when the client reads the information from a field
descriptor and tries to store it into the fields structure.

FIX:
====
Check for the appropriate fields data and error out if it
is not present.

RB: 22440