BUG#30192182
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

Calls to execNODE_FAILREP in SafeCounter expected a node bitmask in
signal object. This is no longer present there, in addition it can
be destroyed by call to this function so the second call cannot be
trusted.

Solved by adding a new parameter to the function where the node bitmask
is carried.

Removed some overly talky debug statements.
Fixed some downgrade code where initialisation of NODE_FAILREP bitmask
was in wrong place.