Bug #30192171 - WL#10358: SIG6 IN TEMPTABLE::HANDLER::POSITION() AT SRC/HANDLER.CC [noclose]
When deciding whether to use addon fields or not, we need to make sure
we have the correct sort order set up; using_addon_fields() would depend
on data from init_for_filesort(), and could thus make a wrong choice if
called too early.

Ironically, this means that if you activated EXPLAIN printing of a crashing
query, it would call using_addon_fields() from DebugString(), and avoid
crashing because it made a more lucky choice!

This was only seen on WL #10358, but is not actually related to it.
The test case depends on that worklog, though.

Change-Id: Idf382cff318e1705c7cb82eda0295b1ee0b1ae35