Bug #30192171 - WL#10358: SIG6 IN TEMPTABLE::HANDLER::POSITION() AT SRC/HANDLER.CC [noclose]
If we have filesort both for duplicate removal _and_ ORDER BY (ie., two
separate sorts after each other), the former needs to be the one to decide
whether we can stream rows or not; otherwise, we could end up in a situation
where we'd set up streaming even though filesort expected to refer to the
materialized rows later.

This was only seen on WL #10358, but is not actually related to it.
The test case depends on that worklog, though.

Change-Id: I5cf997e0b545c33a48e77b98569ed45d1eeac347