Bug#29906966: Failed assertion when trying to create a column with literal
Worklog #12358 added a few asserts on which basic literals represented
as Item_func objects that are allowed to be used as default values
for table columns. However, the literals for TIMESTAMP, DATE and TIME
were omitted, which oddly enough are not allowed as default values.

This bug fix augments the check by validating that the type of
the basic const literal is integer (TRUE and FALSE adhere to this),
and makes sure that TIMESTAMP, DATE and TIME literals are rejected,
as was the case before the regression.

Change-Id: If3b25779ec6399b1835263497c9c5aa572f87d94