Bug#27492656: MTR IGNORES THE INIT_FILE OPTION WITH --BOOTSTRAP DURING               SERVER INITIALIZATION
Post push fix: If init-file is not passed on the command line, by default
it was set to the build directory unintentionally, which could not be opened
as a file on Windows. On Linux, if the file to be opened turned out to
be a directory, it would not throw an error. Now, the init-file stays
undefined if it is not used on the command line.

Change-Id: I01546c0bc6b699fb7ebaf097e40107a38dcd2b12